refact: move action buttons to comp file
