+++ app/dashboard/src/pages/nodes/table/index.tsx
@@ -5,7 +5,6 @@ import {
   AccordionItem,
   AccordionPanel,
   Box,
-  chakra,
   HStack,
   IconButton,
   Table,
@@ -21,8 +20,8 @@ import {
   VStack,
 } from '@chakra-ui/react';
 import classNames from 'classnames';
+import { ActionButtons } from './action-buttons';
 
-import { t } from 'i18next';
 import { FC, Fragment, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { StatusBadge } from 'components/status-badge';
@@ -36,16 +35,6 @@ import {
   handleSort,
 } from 'components/table';
 import { useNodes } from 'stores';
-import { Cog6ToothIcon } from '@heroicons/react/24/outline';
-
-const iconProps = {
-  baseStyle: {
-    w: 4,
-    h: 4,
-  },
-};
-
-const CoreSettingsIcon = chakra(Cog6ToothIcon, iconProps);
 
 type ExpandedIndex = number | number[];
 
@@ -403,43 +392,5 @@ export const NodesTable: FC<NodesTableProps> = (props) => {
       </Table>
       <Pagination filters={filters} total={total} onFilterChange={onFilterChange} />
     </Box>
-  );
-};
-
-type ActionButtonsProps = {};
-
-const ActionButtons: FC<ActionButtonsProps> = () => {
-  return (
-    <HStack
-      justifyContent="flex-end"
-      onClick={(e) => {
-        e.preventDefault();
-        e.stopPropagation();
-      }}
-    >
-      <Tooltip
-        label={
-          t('nodesTable.coreSettings')
-        }
-        placement="top"
-      >
-        <IconButton
-          p="0 !important"
-          aria-label="core settings"
-          bg="transparent"
-          _dark={{
-            _hover: {
-              bg: 'gray.700',
-            },
-          }}
-          size={{
-            base: 'sm',
-            md: 'md',
-          }}
-        >
-          <CoreSettingsIcon />
-        </IconButton>
-      </Tooltip>
-    </HStack >
   );
 };
